+++ scripts/module.js
@@ -2,6 +2,7 @@ import { MODULE_ID, MODULE_TITLE } from "./constants.js";
 import { exposeApi } from "./api.js";
 import { registerArmorProfileHooks, registerPacsEquipmentSlots } from "./armor-profile.js";
 import { patchD35EAttackRolls } from "./d35e-integration.js";
+import { registerHelmetSkillPenaltyHook } from "./helmet.js";
 import { registerSettings } from "./settings.js";
 import { registerUiHooks } from "./ui.js";
 
@@ -20,6 +21,7 @@ Hooks.once("ready", async () => {
   }
   registerPacsEquipmentSlots();
   await patchD35EAttackRolls();
+  registerHelmetSkillPenaltyHook();
   registerArmorProfileHooks();
   registerUiHooks();
   console.info(`${MODULE_TITLE} | Ready.`);